materialize modifications

+++ src/main/java/com/codeup/blog/RollDiceController.java
@@ -4,7 +4,6 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.ResponseBody;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -15,14 +14,14 @@ public class RollDiceController {
 
     @GetMapping("/roll-dice")
     public String hi(Model model) {
-        double randomNum = 0;
+        int randomNum = 0;
         model.addAttribute("randomNum", randomNum);
         return "roll-dice";
     }
 
     @GetMapping("/roll-dice/{urlNum}")
-    public String rollNum(@PathVariable double urlNum, Model model) {
-        double randomNum = (long) (Math.random() * 6) + 1;
+    public String rollNum(@PathVariable int urlNum, Model model) {
+        int randomNum = (int) (Math.random() * 6) + 1;
         boolean correctPrediction = false;
         model.addAttribute("randomNum", randomNum);
         model.addAttribute("urlNum", urlNum);
